Maze App fully functional
Doesn't seem to be any bugs, the maze is completely functional (you can reach every room). I just haven't had time to write descriptions/add items for every single one.

The global app data (like current inventory/status of each room) is also successfully reset when you reach the end of the maze.

I planned for adding a lock/key system ,as well as a maze shop! But that's beyond the scope of the assignment, so I just left the items there.